refactor: :recycle: code formatting for better developer experience
