+++ package.json
@@ -10,7 +10,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@faker-js/faker": "^8.4.1",
     "@tanstack/react-query": "^5.28.4",
     "@tanstack/react-router": "^1.20.0",
     "@tanstack/react-table": "^8.13.2",
