+++ src/types/IconButtonProps.ts
@@ -1,0 +1,5 @@
+export default interface IconButtonProps {
+  onClick: () => void;
+  disabled?: boolean;
+  children: React.ReactNode;
+}
