+++ tsconfig.json
@@ -14,6 +14,7 @@
       "hooks/*":["./src/hooks/*"],
       "services/*":["./src/services/*"],
       "store/*":["./src/store/*"],
+      "types/*":["./src/types/*"],
       "utils/*":["./src/utils/*"],
     },
